update vendor/libxev to retry on EINTR for io_uring
Fixes #290

Previously, our io_uring loops were exiting on EINTR. EINTR is sent for
various reasons but particularly when a process is stopped such as for
sleeping. We should not exit these loops.

Longer term we should be more resilient to detecting that our renderer
and/or IO thread as died and behave in some better way but this
addresses a major issue in the mean time.